+++ src/main/java/dk/kb/storage/api/v1/impl/DsStorageApiServiceImpl.java
@@ -8,6 +8,7 @@ import dk.kb.storage.model.v1.MappingDto;
 import dk.kb.storage.model.v1.OriginCountDto;
 import dk.kb.storage.model.v1.OriginDto;
 import dk.kb.storage.model.v1.RecordTypeDto;
+import dk.kb.storage.model.v1.RecordsCountDto;
 import dk.kb.util.Pair;
 import dk.kb.util.webservice.ImplBase;
 import dk.kb.util.webservice.stream.ExportWriter;
@@ -215,6 +216,16 @@ public class DsStorageApiServiceImpl extends ImplBase implements DsStorageApi {
     }
 
     @Override
+    public RecordsCountDto touchRecord(String recordId) {
+        try {
+            log.debug("touchRecord(id='{}') called with call details: {}", recordId, getCallDetails());
+            return DsStorageFacade.touchRecord(recordId);
+        } catch (Exception e) {
+            throw handleException(e);
+        }
+    }
+
+    @Override
     public DsRecordDto getRecord(String id, Boolean includeLocalTree) {
         try {
             log.debug("getRecord(id='{}') called with call details: {}", id, getCallDetails());
@@ -241,7 +252,7 @@ public class DsStorageApiServiceImpl extends ImplBase implements DsStorageApi {
     */
      
     @Override
-    public Integer markRecordForDelete(String id) {
+    public RecordsCountDto markRecordForDelete(String id) {
         try {
             log.debug("markRecordForDelete(id='{}') called with call details: {}", id, getCallDetails());
             return DsStorageFacade.markRecordForDelete(id);
@@ -251,7 +262,7 @@ public class DsStorageApiServiceImpl extends ImplBase implements DsStorageApi {
     }
 
     @Override
-    public Integer deleteMarkedForDelete(String origin) {
+    public RecordsCountDto deleteMarkedForDelete(String origin) {
         try {
             log.debug("deleteMarkedForDelete(origin'{}') called with call details: {}",
                       origin, getCallDetails());
@@ -272,7 +283,7 @@ public class DsStorageApiServiceImpl extends ImplBase implements DsStorageApi {
     }
 
     @Override
-    public Integer deleteRecordsForOrigin(String origin, Long mTimeFrom, Long mTimeTo) {
+    public RecordsCountDto deleteRecordsForOrigin(String origin, Long mTimeFrom, Long mTimeTo) {
         try {
 	    log.debug("deleteRecordsForOrigin() called with call details: {}", getCallDetails());
 	    return DsStorageFacade.deleteRecordsForOrigin(origin,mTimeFrom,mTimeTo);
@@ -322,7 +333,7 @@ public class DsStorageApiServiceImpl extends ImplBase implements DsStorageApi {
         }        
     }
     @Override
-    public Integer updateKalturaIdForRecords() {
+    public RecordsCountDto updateKalturaIdForRecords() {
         log.debug("updateKalturaIdForRecords() called with call details: {}", getCallDetails());
         return DsStorageFacade.updateKalturaIdForRecords();        
     }