+++ src/main/java/dk/kb/storage/util/DsStorageClient.java
@@ -14,26 +14,30 @@
  */
 package dk.kb.storage.util;
 
-import dk.kb.storage.client.v1.DsStorageApi;
-import dk.kb.storage.invoker.v1.ApiClient;
-import dk.kb.storage.invoker.v1.ApiException;
-import dk.kb.storage.invoker.v1.Configuration;
 import dk.kb.storage.model.v1.DsRecordDto;
 import dk.kb.storage.model.v1.DsRecordMinimalDto;
 import dk.kb.storage.model.v1.MappingDto;
+import dk.kb.storage.model.v1.OriginCountDto;
+import dk.kb.storage.model.v1.OriginDto;
 import dk.kb.storage.model.v1.RecordTypeDto;
+import dk.kb.storage.model.v1.RecordsCountDto;
+import dk.kb.util.webservice.Service2ServiceRequest;
 import dk.kb.util.webservice.exception.InternalServiceException;
+import dk.kb.util.webservice.exception.ServiceException;
 import dk.kb.util.webservice.stream.ContinuationInputStream;
 import dk.kb.util.webservice.stream.ContinuationStream;
-import org.apache.http.client.utils.URIBuilder;
+
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
+
+import org.apache.hc.core5.net.URIBuilder;
 
 /**
  * Client for the service. Intended for use by other projects that calls this service.
@@ -44,12 +48,13 @@ import java.util.Locale;
  * The client is Thread safe and handles parallel requests independently.
  * It is recommended to persist the client and to re-use it between calls.
  */
-public class DsStorageClient extends DsStorageApi {
+public class DsStorageClient {
     private static final Logger log = LoggerFactory.getLogger(DsStorageClient.class);
+    private final static String CLIENT_URL_EXCEPTION="The client url was not constructed correct";
     private final String serviceURI;
 
     public static final String STORAGE_SERVER_URL_KEY = ".storage.url";
-    
+
     /**
      * Creates a client for the remote ds-storage service.
      * <p>
@@ -63,44 +68,193 @@ public class DsStorageClient extends DsStorageApi {
      */
     @SuppressWarnings("JavadocLinkAsPlainText")
     public DsStorageClient(String serviceURI) {
-        super(createClient(serviceURI));
         this.serviceURI = serviceURI;
         log.info("Created OpenAPI client for '{}'", serviceURI);
     }
 
-    
     /**
-     * <p>
-     * If the mapping does not exist a new entry will be created in the mapping table.<br>
-     * The referenceid can not be null, but kalturaId can be null.<br>
-     * If the mapping already exist for the referenceid, the kalturaId value will be updated
-     * </p>
-     * 
-     * @param mapping The mapping entry to be created or updated
-     * 
+     * Retrieve a list of configured origins with their respective update strategy
+     * This endpoint delivers a list of all configured origins. An origin defines which collection data comes from. This could for instance be the Radio &amp; TV collection at The Royal Danish Library, which has the origin defined as &#39;ds.radiotv&#39;. The update strategy defines how data from the specific origin is updated, when a record is added, modified or deleted. 
+     * @return List&lt;OriginDto&gt;
+     * @throws ServiceException if fails to make API call
      */
-    public void updateMappings(MappingDto mapping) throws ApiException {               
-       super.mappingPost(mapping);                       
+    public List<OriginDto> getOriginConfiguration () throws ServiceException{       
+        try {
+            URI uri = new URIBuilder(serviceURI)                    
+                    .appendPathSegments("origin","config")                                                                
+                    .build();
+            return Service2ServiceRequest.httpCallWithOAuthTokenAsDtoList(uri,"GET",new OriginDto(),null);              
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);               
+         }          
+
     }
 
     /**
-     * This method is deprecated. See: {@link #getDsRecordsMinimalModifiedAfterStream(String, int, long)}
-     * Get a list of records having a referenceId after a given lastModified time
-     * Extract a list of records with a given batch size by origin and mTime larger than input.
-     * The records will only have the id, mTime, referenceId and kalturaId fields set.
-     * <p/>
-     *  
-     *  @param origin The Origin to extract records from
-     *  @param batchSize How many records to fetch.
-     *  @param mTimeFrom Only retrieve records after mTimeFrom
+     * Show amount of records in each origin
      * 
-     * @throws ApiException  
-     * 
+     * @return List&lt;OriginCountDto&gt;
+     * @throws ServiceException if fails to make API call
      */
-    public List<DsRecordMinimalDto> getDsRecordsReferenceIdModifiedAfter(String origin,int batchSize,long mTimeFrom) throws ApiException {
-        return super.getMinimalRecords(origin, batchSize, mTimeFrom);
+    public List<OriginCountDto> getOriginStatistics() throws ServiceException {
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("origin","stats")                                                                                   
+                    .build();
+            return Service2ServiceRequest.httpCallWithOAuthTokenAsDtoList(uri,"GET",new OriginCountDto(),null);              
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);               
+         }                        
     }
+
+  
+
     
+    /**
+     * Read a specific record by ID.
+     * Extract a specific record by ID. Parent and children recordIds will also be include. If setting includeLocalTree&#x3D;true the local recordTree with parent record and children records will also be loaded as objects.   A record marked with delete flag will also be returned. If a record is not found in ds-storage, the endpoint will throw an exception. 
+     * @param id Record ID (required)
+     * @param includeLocalTree Also load parent and direct children as objects (optional, default to false)
+     * @return DsRecordDto
+     * @throws ServiceException if fails to make API call
+     */
+    public DsRecordDto getRecord(String id, Boolean includeLocalTree) throws ServiceException{       
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("record",id) //id is part of path                                             
+                    .addParameter("includeLocalTree",""+includeLocalTree)               
+                    .build();
+            return Service2ServiceRequest.httpCallWithOAuthToken(uri,"GET",new DsRecordDto(),null);              
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);               
+         }           
+    }
+
+    public RecordsCountDto touchRecord(String recordId){
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("record", "touch", recordId)
+                    .build();
+            return Service2ServiceRequest.httpCallWithOAuthToken(uri,"POST", new RecordsCountDto(),null);
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url: '{}'", e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);
+        }
+    }
+
+
+    /**
+     * Create a new record or update an existing record.
+     * A record can have a single optional parent and may have multiple children
+     * @param dsRecordDto  (optional)
+     * @throws ServiceException if fails to make API call
+     */
+    public void recordPost(DsRecordDto dsRecordDto) throws ServiceException {
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("record")
+                    .build();
+            Service2ServiceRequest.httpCallWithOAuthToken(uri,"POST", null, dsRecordDto);              
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);               
+         }         
+    }
+
+    /**
+     * Mark a record with delete flag.
+     * This will not delete the record in the database but only mark it as deleted. 
+     * @param id Record ID (required)
+     * @return Integer
+     * @throws ServiceException if fails to make API call
+     */
+    public RecordsCountDto markRecordForDelete (String id) throws ServiceException {
+        try {                          
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("record",id) //id is part of path                                                                                  
+                    .build();
+            return Service2ServiceRequest.httpCallWithOAuthToken(uri,"DELETE",new RecordsCountDto(),null);              
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);               
+        }                    
+    }
+
+    /**
+     * Create a new mapping or update a mapping .
+     * Create a new mapping or update mapping if referenceId exists. Each record with a stream will have a referenceId (file-id) and needs to be mapped to the KalturaId
+     * @param mappingDto  (optional)
+     * @throws ServiceException if fails to make API call
+     */
+    public void mappingPost(MappingDto mapping) throws ServiceException {               
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                     .appendPathSegments("mapping")                                                               
+                     .build();
+            Service2ServiceRequest.httpCallWithOAuthToken(uri,"POST", null, mapping);              
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);               
+        }                               
+    }
+
+    /**
+     * Get mapping for a specific referenceId
+     * Get entry from the mapping table for the referenceId. If the entry is not found null will be returned. It is not guarantees the entry if it exists, will have the kalturaId set yet.
+     * @param referenceId  (required)
+     * @return MappingDto
+     * @throws ServiceException if fails to make API call
+     */
+    public MappingDto getMapping(String referenceId) throws ServiceException {        
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("mapping")                                                                
+                    .addParameter("referenceId",referenceId)                       
+                    .build();
+            return Service2ServiceRequest.httpCallWithOAuthToken(uri,"GET", new MappingDto(), null);              
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);               
+         }                               
+    }
+        
+    /**
+     * Get a list of minimal records having a referenceId after a given lastModified time
+     * Extract a list of records with a given batch size by origin and mTime larger than input. 
+     * The records will only have the id, mTime, referenceId and kalturaId fields. This means that no actual data can be retrieved through this endpoint. It can however be used  for operations where the data from the record isn&#39;t needed. Such as updating Kaltura IDs for records, which is done with referenceId and kalturaId only. 
+     * @param origin The origin to extract records for (required)
+     * @param maxRecords Number of records to extract. (required)
+     * @param mTime Only extract records after this mTime. (optional, default to 0l)
+     * @return List&lt;DsRecordMinimalDto&gt;
+     * @throws ServiceException if fails to make API call
+     */
+    public List<DsRecordMinimalDto> getMinimalRecords (String origin, Integer maxRecords, Long mTime) throws ServiceException {   
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("records","minimal")                                                               
+                    .addParameter("origin",origin)
+                    .addParameter("maxRecords",""+maxRecords)
+                    .addParameter("mTime",""+mTime)                    
+                    .build();
+            return Service2ServiceRequest.httpCallWithOAuthToken(uri,"GET", new ArrayList<DsRecordMinimalDto>(), null);              
+        }
+        catch (URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);               
+         }                                            
+    }
+
     /**
      * Call the remote ds-storage {@link #getRecordsModifiedAfter} and return the response in the form of a Stream 
      * of records.
@@ -160,19 +314,16 @@ public class DsStorageClient extends DsStorageApi {
             throws IOException {
         URI uri;
         try {
-            uri = new URIBuilder(serviceURI + "records")
-                    // setPath overwrites paths given in serviceURI
-                    // .setPath("records")
+            uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("records")                                
                     .addParameter("origin", origin)
                     .addParameter("mTime", Long.toString(mTime == null ? 0L : mTime))
                     .addParameter("maxRecords", Long.toString(maxRecords == null ? 10 : maxRecords))
                     .build();
-        } catch (URISyntaxException e) {
-            String message = String.format(Locale.ROOT,
-                    "getRecordsModifiedAfterJSON(origin='%s', mTime=%d, maxRecords=%d): Unable to construct URI",
-                    origin, mTime, maxRecords);
-            log.warn(message, e);
-            throw new InternalServiceException(message);
+        }
+        catch (URISyntaxException e) {
+                log.error("Invalid url:"+e.getMessage());
+                throw new InternalServiceException(CLIENT_URL_EXCEPTION);                           
         }
 
         log.debug("Opening streaming connection to '{}'", uri);
@@ -196,25 +347,22 @@ public class DsStorageClient extends DsStorageApi {
             String origin, RecordTypeDto recordType, Long mTime, Long maxRecords) throws IOException {
         URI uri;
         try {
-            uri = new URIBuilder(serviceURI + "records")
-                    // setPath overwrites paths given in serviceURI
-                    //.setPath("records")
+
+            uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("records")
                     .addParameter("origin", origin)
                     .addParameter("recordType", recordType.toString())
                     .addParameter("mTime", Long.toString(mTime == null ? 0L : mTime))
                     .addParameter("maxRecords", Long.toString(maxRecords == null ? 10 : maxRecords))
                     .build();
-        } catch (URISyntaxException e) {
-            String message = String.format(Locale.ROOT,
-                    "getRecordsModifiedAfterLocalTreeJSON(origin='%s', recordType='%s', mTime=%d, maxRecords=%d): " +
-                            "Unable to construct URI",
-                    origin, recordType, mTime, maxRecords);
-            log.warn(message, e);
-            throw new InternalServiceException(message);
+        }
+        catch (URISyntaxException e) {
+                log.error("Invalid url:"+e.getMessage());
+                throw new InternalServiceException(CLIENT_URL_EXCEPTION);                      
         }
 
-        log.debug("Opening streaming connection to '{}'", uri);
-        return ContinuationInputStream.from(uri, Long::valueOf);
+        log.debug("Opening streaming connection to '{}'", uri);      
+        return ContinuationInputStream.from(uri, Long::valueOf);       
     }
 
     /**
@@ -253,26 +401,22 @@ public class DsStorageClient extends DsStorageApi {
             throws IOException {
         URI uri;
         try {
-            uri = new URIBuilder(serviceURI + "records/minimal")
-                    // setPath overwrites paths given in serviceURI
-                    // .setPath("records")
+            uri = new URIBuilder(serviceURI)                    
+                    .appendPathSegments("records","minimal")                                        
                     .addParameter("origin", origin)
                     .addParameter("mTime", Long.toString(mTime == null ? 0L : mTime))
                     .addParameter("maxRecords", Long.toString(maxRecords == null ? 10 : maxRecords))
                     .build();
         } catch (URISyntaxException e) {
-            String message = String.format(Locale.ROOT,
-                    "getMinimalRecordsModifiedAfterJSON(origin='%s', mTime=%d, maxRecords=%d): Unable to construct URI",
-                    origin, mTime, maxRecords);
-            log.warn(message, e);
-            throw new InternalServiceException(message);
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);
         }
 
         log.debug("Opening streaming connection to '{}'", uri);
         return ContinuationInputStream.from(uri, Long::valueOf);
     }
 
-    
+
     /**
      * Update the referenceId for a record <br>
      * The referenceId is an id in the external system for the record. <br>
@@ -281,27 +425,68 @@ public class DsStorageClient extends DsStorageApi {
      *  @param recordId of the record to update referenceId for
      *  @param referenceId The referenceId to set for the record
      * 
-     * @throws ApiException  
+     * @throws ServiceException  
      * 
      */
-    public void updateReferenceIdForRecord(String recordId,String referenceId) throws ApiException {
-        super.updateReferenceIdForRecord(recordId, referenceId);
+    public void updateReferenceIdForRecord(String recordId,String referenceId) throws ServiceException{       
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    
+                    .appendPathSegments("record","updateReferenceId")                                                               
+                    .addParameter("recordId",""+recordId)               
+                    .addParameter("referenceId",""+referenceId)
+                    .build();            
+            Service2ServiceRequest.httpCallWithOAuthToken(uri,"POST", null,null);              
+        }
+        catch(URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);
+        }                        
+
+    }
+
+    /**
+     * Update a record with the Kaltura id. 
+     * Update a record with the Kaltura id. The record was uploaded to Kaltura with the referenceId as metadata. Knowing the kalturaId is important to find the record(stream) in Kaltura later for update or deletetion etc.
+     * @param referenceId  (required)
+     * @param kalturaId  (required)
+     * @throws ServiceException if fails to make API call
+     */
+    public void updateKalturaIdForRecord (String referenceId, String kalturaId) throws ServiceException {
+
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("record","updateKalturaId")                                                                                 
+                    .addParameter("referenceId",referenceId)               
+                    .addParameter("kalturaId",kalturaId)
+                    .build();            
+            Service2ServiceRequest.httpCallWithOAuthToken(uri,"POST", null,null);              
+        }
+        catch(URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);
+        }                                      
     }
     
     /**
-     * Deconstruct the given URI and use the components to create an ApiClient.
-     * @param serviceURIString a URI to a service.
-     * @return an ApiClient constructed from the serviceURIString.
+     * Update all records having a referenceIf with the matching Kaltura, if the mapping can be found in the mapping table
+     * Update all records having a referenceId with the matching kalturaId, if the mapping can be found in the mapping table.  If many records needs to be updated this can take some time. Estimated 15 minutes for 1M records. It is possible to update in batches if the mapping table is also updated in batches.                 
+     * @return Integer
+     * @throws ApiException if fails to make API call
      */
-    private static ApiClient createClient(String serviceURIString) {
-        log.debug("Creating OpenAPI client with URI '{}'", serviceURIString);
-
-        URI serviceURI = URI.create(serviceURIString);
-        // No mechanism for just providing the full URI. We have to deconstruct it
-        return Configuration.getDefaultApiClient().
-                setScheme(serviceURI.getScheme()).
-                setHost(serviceURI.getHost()).
-                setPort(serviceURI.getPort()).
-                setBasePath(serviceURI.getRawPath());
+    public RecordsCountDto updateKalturaIdForRecords() throws ServiceException {
+        try {
+            URI uri = new URIBuilder(serviceURI)
+                    .appendPathSegments("record","updateKalturaId")                                                                                 
+                    .build();            
+           return Service2ServiceRequest.httpCallWithOAuthToken(uri,"POST", new RecordsCountDto(),null);              
+        }
+        catch(URISyntaxException e) {
+            log.error("Invalid url:"+e.getMessage());
+            throw new InternalServiceException(CLIENT_URL_EXCEPTION);
+        }                             
     }
+       
+
+  
 }